separate parlamint module
